refactor: consolidate duplicated code and remove dead dependencies
- Consolidate checksum algorithms into single calculateChecksum in
  dataUtils.ts, supporting all protocol algorithms (CRC16_MODBUS,
  CRC16_CCITT, LRC). Remove duplicate implementations from
  messageParser.ts and messageBuilder.ts.
- Extract duplicated line ending switch blocks into appendLineEnding
  utility, replacing 4 identical switch statements across InputPanel,
  useCommandExecution, and MainWorkspace.
- Remove unused lazy_static dependency from Cargo.toml.
- Delete orphaned useFilteredItems.ts hook (zero imports).

Closes #85, #86, #90, #99

via [Happy](https://happy.engineering)
