+++ src-tauri/Cargo.toml
@@ -21,8 +21,6 @@ serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.149"
 
 
-lazy_static = "1.5.0"
-
 tauri-plugin-log = "^2.8.0"
 # tauri-plugin-devtools = "2.0.0-rc"
 tauri-plugin-fs = "^2.4.5"
