+++ src/lib/dataUtils.ts
@@ -1,41 +1,98 @@
-import { ChecksumAlgorithm, TextEncoding } from "../types";
+import { ChecksumAlgorithm, LineEnding, TextEncoding } from "../types";
+
+/**
+ * Supported checksum algorithm names (superset of base + protocol algorithms)
+ */
+type SupportedChecksumAlgorithm =
+  | ChecksumAlgorithm
+  | "CRC16_MODBUS"
+  | "CRC16_CCITT"
+  | "LRC";
 
 export const calculateChecksum = (
   data: Uint8Array,
-  algo: ChecksumAlgorithm,
+  algo: SupportedChecksumAlgorithm,
 ): Uint8Array => {
-  if (algo === "NONE") return new Uint8Array(0);
+  switch (algo) {
+    case "NONE":
+      return new Uint8Array(0);
 
-  if (algo === "MOD256") {
-    let sum = 0;
-    for (const b of data) sum = (sum + b) % 256;
-    return new Uint8Array([sum]);
-  }
+    case "MOD256": {
+      let sum = 0;
+      for (const b of data) sum = (sum + b) % 256;
+      return new Uint8Array([sum]);
+    }
 
-  if (algo === "XOR") {
-    let xor = 0;
-    for (const b of data) xor ^= b;
-    return new Uint8Array([xor]);
-  }
+    case "XOR": {
+      let xor = 0;
+      for (const b of data) xor ^= b;
+      return new Uint8Array([xor]);
+    }
 
-  if (algo === "CRC16") {
-    // CRC-16-MODBUS (Polynomial 0x8005, initial 0xFFFF, reversed)
-    let crc = 0xffff;
-    for (let i = 0; i < data.length; i++) {
-      crc ^= data[i];
-      for (let j = 0; j < 8; j++) {
-        if ((crc & 1) !== 0) {
-          crc = (crc >> 1) ^ 0xa001;
-        } else {
-          crc = crc >> 1;
+    case "CRC16":
+    case "CRC16_MODBUS": {
+      // CRC-16-MODBUS (Polynomial 0x8005, initial 0xFFFF, reversed)
+      let crc = 0xffff;
+      for (let i = 0; i < data.length; i++) {
+        crc ^= data[i];
+        for (let j = 0; j < 8; j++) {
+          if ((crc & 1) !== 0) {
+            crc = (crc >> 1) ^ 0xa001;
+          } else {
+            crc = crc >> 1;
+          }
         }
       }
+      // Modbus sends Low Byte first, High Byte second
+      return new Uint8Array([crc & 0xff, (crc >> 8) & 0xff]);
     }
-    // Modbus sends Low Byte first, High Byte second
-    return new Uint8Array([crc & 0xff, (crc >> 8) & 0xff]);
-  }
 
-  return new Uint8Array(0);
+    case "CRC16_CCITT": {
+      // CRC-16-CCITT (Polynomial 0x1021, initial 0xFFFF)
+      let crc = 0xffff;
+      for (let i = 0; i < data.length; i++) {
+        crc ^= data[i] << 8;
+        for (let j = 0; j < 8; j++) {
+          if ((crc & 0x8000) !== 0) {
+            crc = ((crc << 1) ^ 0x1021) & 0xffff;
+          } else {
+            crc = (crc << 1) & 0xffff;
+          }
+        }
+      }
+      return new Uint8Array([(crc >> 8) & 0xff, crc & 0xff]);
+    }
+
+    case "LRC": {
+      // Longitudinal Redundancy Check (two's complement of sum)
+      let sum = 0;
+      for (const b of data) sum = (sum + b) & 0xff;
+      const lrc = (~sum + 1) & 0xff;
+      return new Uint8Array([lrc]);
+    }
+
+    default:
+      return new Uint8Array(0);
+  }
+};
+
+/**
+ * Append the appropriate line ending characters to a string
+ */
+export const appendLineEnding = (
+  data: string,
+  lineEnding: LineEnding,
+): string => {
+  switch (lineEnding) {
+    case "LF":
+      return data + "\n";
+    case "CR":
+      return data + "\r";
+    case "CRLF":
+      return data + "\r\n";
+    default:
+      return data;
+  }
 };
 
 export const encodeText = (
