+++ src/lib/messageParser.ts
@@ -18,10 +18,10 @@ import type {
   MessageElement,
   ByteOrder,
   DataType,
-  ChecksumAlgorithm,
   ResponsePattern,
 } from "./protocolTypes";
 import { executeSandboxedScript } from "./sandboxedScripting";
+import { calculateChecksum } from "./dataUtils";
 
 // ============================================================================
 // TYPES
@@ -155,72 +155,6 @@ function getDataTypeSize(dataType: DataType): number {
       return 8;
     default:
       return 1;
-  }
-}
-
-/**
- * Calculate checksum for validation
- */
-function calculateChecksum(
-  data: Uint8Array,
-  algorithm: ChecksumAlgorithm,
-): Uint8Array {
-  switch (algorithm) {
-    case "NONE":
-      return new Uint8Array(0);
-
-    case "MOD256": {
-      let sum = 0;
-      for (const b of data) sum = (sum + b) % 256;
-      return new Uint8Array([sum]);
-    }
-
-    case "XOR": {
-      let xor = 0;
-      for (const b of data) xor ^= b;
-      return new Uint8Array([xor]);
-    }
-
-    case "CRC16":
-    case "CRC16_MODBUS": {
-      let crc = 0xffff;
-      for (let i = 0; i < data.length; i++) {
-        crc ^= data[i];
-        for (let j = 0; j < 8; j++) {
-          if ((crc & 1) !== 0) {
-            crc = (crc >> 1) ^ 0xa001;
-          } else {
-            crc = crc >> 1;
-          }
-        }
-      }
-      return new Uint8Array([crc & 0xff, (crc >> 8) & 0xff]);
-    }
-
-    case "CRC16_CCITT": {
-      let crc = 0xffff;
-      for (let i = 0; i < data.length; i++) {
-        crc ^= data[i] << 8;
-        for (let j = 0; j < 8; j++) {
-          if ((crc & 0x8000) !== 0) {
-            crc = ((crc << 1) ^ 0x1021) & 0xffff;
-          } else {
-            crc = (crc << 1) & 0xffff;
-          }
-        }
-      }
-      return new Uint8Array([(crc >> 8) & 0xff, crc & 0xff]);
-    }
-
-    case "LRC": {
-      let sum = 0;
-      for (const b of data) sum = (sum + b) & 0xff;
-      const lrc = (~sum + 1) & 0xff;
-      return new Uint8Array([lrc]);
-    }
-
-    default:
-      return new Uint8Array(0);
   }
 }
